% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row}
\alias{add_row}
\title{add_row}
\usage{
add_row(mydf, ...)
}
\arguments{
\item{mydf,}{mylist (a data frame and a list)}
}
\description{
How to elegantly add one row to a data frame, similar to SQL INSERT INTO. Maintain same class as original data frame.
WARNING: THIS IS STILL A WORK IN PROGRESS!!!
}
\examples{
# Idea from Hadley Wickham
 add_row(mtcars, cyl = 4, disp = 7)
}
\keyword{dataframes}

