% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_from_list.R
\name{add_row_from_list}
\alias{add_row_from_list}
\title{add_row_from_list Add one row of data to a data frame from a list}
\usage{
add_row_from_list(mydf, mylist)
}
\arguments{
\item{mydf}{A data frame to which you'd like to add a row}

\item{mylist}{A list holding a single row of data,
in the exact order of columns in the original data frame.}
}
\description{
Easily add one row to a data frame when columns are not all of one class. Data should be in a list format, one row only, in same order as columns in the original data frame. Any missing values must be explicitly included as NAs. Similar to SQL INSERT INTO. Maintains same class as original data frame.
}
\details{
WARNING: THIS IS STILL A WORK IN PROGRESS!!!
}
\examples{
# Idea from Jennifer Bryan
add_row_from_list(iris, list(5.1, 3.5, 1.4, 0.2, "setosa"))


}
\keyword{dataframes}

