% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strings.R
\name{copy_textfile_to_clipboard}
\alias{copy_textfile_to_clipboard}
\title{copy_textfile_to_clipboard}
\usage{
copy_textfile_to_clipboard(mytextfile)
}
\arguments{
\item{mytextfile}{Path to a text file as a character string}
}
\description{
This function takes the path to a textfile as a character string and reads it into the Windows or Mac clipboards only.
}
\examples{
myfile <- "data/textfile.txt"
copy_textfile_to_clipboard(myfile)
}
\keyword{clipboard}

