% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_project_file.R
\name{edit_project_list}
\alias{edit_project_list}
\title{Open file to edit RStudio project list}
\usage{
edit_project_list(windrive = "C")
}
\arguments{
\item{windrive}{For Windows only, drive where RStudio is installed. Defaults to C.}
}
\value{
action opens file
}
\description{
Opens file for editing the RStudio drop-down project list so individual projects can be removed from the list.
}
\examples{
\dontrun{
edit_project_list()
}
}
