% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elect_pcts_by_row.R
\name{elec_pcts_by_row}
\alias{elec_pcts_by_row}
\title{elec_pcts_by_row}
\usage{
elec_pcts_by_row(df, mycols, totalcol = "Total")
}
\arguments{
\item{df}{data frame that contains at least one numeric column.}

\item{mycols}{character strings - names of columns with numerical data for which you'd like to calculate percentages across rows.}

\item{totalcol}{character string - name of column that has vote (or other) totals.}
}
\value{
data frame
}
\description{
This function lets you select columns in a data frame with raw numbers and will add columns with percent data by row. This is particular useful for election results where each row contains data by precinct, city, etc. and each column is a different candidate, ballot question yes/no, or similar construct.
}
\details{
This is a modification of the pct function from the caroline package by David Schruth.
}
\examples{
City <- c("Cityville", "Townville", "Placeville")
Yes <- c(250, 100, 150)
No <- c(200, 200, 300)
mydf <- data.frame(City, Yes, No)
mydf$Total <- mydf$Yes + mydf$No

}
