% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electionutils.R
\name{findwinner}
\alias{findwinner}
\title{findwinner}
\usage{
findwinner(filename, datacolstart, datacolstop, exportcsv = TRUE)
}
\arguments{
\item{filename}{string name of file, either spreadsheet or csv, including extension and file path if not in your working directory}

\item{datacolstart}{integer number of the column where the results data starts. If data starts in column B of an Excel spreadsheet, for example, datacolstart should be 2.}

\item{datacolstop}{integer number of the column where the results data ends}

\item{exportcsv}{Boolean If you want results saved to a csv file (name is same as filename_winners.csv)}
}
\value{
dataframe
}
\description{
This function finds the top vote-getter in a csv or Excel election results file - it looks for the largest number in each row and creates a new column with the name of the winner (more specifically, the column name with the maximum value for each row). Requires rio package.
}
\details{
Results file must have candidate results in adjoining columns, in a format with each candidate having their own column of results. findwinner function arguments: filename (string), datacolstart (number of data column where results data starts; 2 for column B in Excel, for example), datacolstop (number of the last dat…
}
\examples{
\dontrun{
mydata <- findwinner("results.xlsx", 2, 4)
}

}

