% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpres.R
\name{getparty}
\alias{getparty}
\title{Get US President's Polical Party by Date}
\usage{
getparty(thedate)
}
\arguments{
\item{thedate}{Any date on or after 1945-04-12 as an R Date object.}
}
\description{
Get the US president's on a given date starting with Harry Truman (D) on 1945-04-12.
}
\details{
Input a date object on or after 1945-04-12 and get the name of the political party of the US president who was in office at that time.
}
\examples{
getparty(as.Date("2000-01-01"))
}
