% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpres.R
\name{getpres}
\alias{getpres}
\title{Get US President by Date}
\usage{
getpres(thedate)
}
\arguments{
\item{thedate}{Any date on or after 1945-04-12 as an R Date object.}
}
\description{
Get the name of US president on a given date starting with Harry Truman.
}
\details{
Input a date object on or after 1945-04-12 and get the name of the US president who was in office at that time.
}
\examples{
getpres(as.Date("2000-01-01"))
}

