% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxn.R
\name{maxn}
\alias{maxn}
\title{maxn}
\usage{
maxn(n)
}
\arguments{
\item{n}{vector of numbers}
}
\value{
top nth value
}
\description{
A function to find the top nth-highest value in a vector
From kohske on StackOverflow
http://stackoverflow.com/questions/10296866/finding-the-column-number-and-value-the-of-second-highest-value-in-a-row/10297025#10297025
}

