% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{mymode}
\alias{mymode}
\title{mymode}
\usage{
mymode(myvector)
}
\arguments{
\item{myvector}{a vector}
}
\description{
Get the statistical mode from a vector of values.
}
\examples{
salaries <- c(32000, 40000, 12000, 12000, 12000, 12000, 12000, 35000, 90000, 100000, 1000000)
themode <- mymode(salaries)
}

