% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numberswithcommas.R
\name{number.with.commas}
\alias{number.with.commas}
\title{numbers.with.commas}
\usage{
number.with.commas(mynumstring)
}
\arguments{
\item{mynumstring}{- one or more character strings such as "3,650"}
}
\description{
Turn a number with commas, which R views as a string, into a numeric class number.
}
\examples{
vect <- c("3,567", "465", "4,000,654")
number.with.commas(vect)
# [1]    3567     465 4000654
}

