% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numberswithcommas.R
\name{number_with_commas}
\alias{number_with_commas}
\title{numbers_with_commas}
\usage{
number_with_commas(mynumstring)
}
\arguments{
\item{mynumstring}{- one or more character strings such as "3,650"}
}
\description{
Turn a number with commas, which R views as a string, into a numeric class number. Will also remove dollar signs.
}
\examples{
vect <- c("3,567", "465", "$4,000,654")
number_with_commas(vect)
# [1]    3567     465 4000654
}

