% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_textfile}
\alias{read_textfile}
\title{read_textfile}
\usage{
read_textfile(filename)
}
\arguments{
\item{filename}{name of a file as a character string in quotation marks}
}
\description{
Read a text file into a variable as a single character string
}
\examples{
mystring <- read_textfile("myfile.txt")

}

