% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamecol.R
\name{renamecol}
\alias{renamecol}
\title{renamecol}
\usage{
renamecol(mydf, oldcolname, newcolname)
}
\arguments{
\item{mydf}{dataframe with column to be renamed}

\item{oldcolname}{character string existing column name}

\item{newcolname}{character string desired new column name}
}
\value{
dataframe
}
\description{
This function is a wrapper for the more cumbersome base R syntax
colnames(mydf)[colnames(mydf) == 'oldcolname'] <- 'newcolname'
or the function
dplyr::rename(mydf,c('foo'='samples'))
}
\examples{
new_mtcars <- renamecol(mtcars, "mpg", "MPG")

}
