% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollingavg.R
\name{rollingavg}
\alias{rollingavg}
\title{rollingavg}
\usage{
rollingavg(myvec, n = 7)
}
\arguments{
\item{myvec}{Numerical vector from which you want to calculate a trailing average}

\item{n}{Number of items to be used for each mean calculation, such as 7 for a 7-day trailing average}
}
\value{
The n-item trailing average of myvec
}
\description{
Generates an n-item trailing average vector from a numerical vector, after first removing any missing values.
}
\examples{
myvector <- c(100, 98, 97.5, 102, 96, 96.5, 97, 94, 92.3, 93, 95, 91, 89, 88, 88.5, 86)
my7dayavg <- rollingavg(myvector)

}

