% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setindex_bygroup}
\alias{setindex_bygroup}
\title{setindex_bygroup}
\usage{
setindex_bygroup(mydf, mycolnum)
}
\arguments{
\item{mydf}{a data frame}

\item{mycolnum}{Number of the id column for your data frame}
}
\description{
Create an auto-indexing column in a data frame grouped by an id column, with the index starting at 1 for each id column group.
}
\examples{
id <- c(1,1,2,2,2,3,4,4,5,5)
cat <- c("A", "B", "A", "B", "C", "B", "C", "D", "A", "E")
mydataframe <- data.frame(id, cat)
mydataframe$step <- setindex_bygroup(mydataframe, 1)

}
