% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaks.R
\name{streaks}
\alias{streaks}
\title{streaks}
\usage{
streaks(myvector, streakitem = "W", nonstreakitem = "L")
}
\arguments{
\item{myvector}{a vector of characters or numbers such as 0 and 1.}

\item{streakitem}{a character or number that stands for the item you're seeking a streak for, such as "W" or 1.}

\item{nonstreakitem}{a character or number that stands for the item that usually interrupts the streak, such as "L" when seeking "W" streaks. There can be other streak-interrupting items not identified, such as "T" for ties.}
}
\description{
Function to find streaks in a vector, typically winning or losing streaks.
Slight modification of function by Jim Albert at https://baseballwithr.wordpress.com/2014/07/07/team-streaks-part-i-2/
}
\examples{
seasonrecord <- c("Win", "Win", "Win", "Loss", "Loss", "Win", "Tie", "Win", "Win", "Win", "Win", "Win", "Win", "Win", "Loss", "Win")
streaks(seasonrecord, "Win", "Loss")
}
\keyword{baseball}
