% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string2varname}
\alias{string2varname}
\title{string2varname}
\usage{
string2varname(x)
}
\arguments{
\item{x}{string}
}
\value{
variable name
}
\description{
This function takes a string and turns it into a variable name.
Code from John Ramey at http://johnramey.net/blog/2010/12/28/converting-a-string-to-a-variable-name-on-the-fly-and-vice-versa-in-r/
}
\examples{

x <- 42
eval(parse(text = "x"))
# [1] 42

}

