% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factors.R
\name{strings.to.ordered.factors}
\alias{strings.to.ordered.factors}
\title{strings.to.ordered.factors}
\usage{
strings.to.ordered.factors(myvector)
}
\description{
Turn a vector of strings into a factors while keeping the existing order of elements in the vector, instead of having them automatically alphabetized.
}
\examples{
mymonths <- c("Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
ordered_months_as_factors <- strings.to.ordered.factors(mymonths)
}
\keyword{factors}
\keyword{ordered}

