% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{table_sorted}
\alias{table_sorted}
\title{table_sorted}
\usage{
table_sorted(myvector)
}
\description{
Get a frequency table from a vector of categorical variables sorted by frequency in descending order
}
\examples{
myvec <- c("white", "blue", "purple", "white", "blue", "blue", "blue", "purple", "yellow", "white")
table_sorted(myvec)
}
