% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strings.R
\name{toinitialcap}
\alias{toinitialcap}
\title{Base R has toupper and tolower - why not toinitialcap?}
\usage{
toinitialcap(mystring)
}
\arguments{
\item{mystring}{(any string)}
}
\description{
This renames the example .simpleCap function in the toupper help file.
}
\examples{
toinitialcap("this is a headline")
}
\keyword{strings}

