% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{unlibrary}
\alias{unlibrary}
\title{unlibrary}
\usage{
unlibrary(pkgname)
}
\arguments{
\item{pkgname}{name of a package as a character string in quotation marks}
}
\description{
Opposite of library - wrapper to detach a package
}
\examples{
unlibrary(devtools)
}
\keyword{environment}

