% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{varname2string}
\alias{varname2string}
\title{varname2string}
\usage{
varname2string(x)
}
\arguments{
\item{x}{variable name}
}
\value{
string
}
\description{
This function takes a variable name and turns it into a string.
Code from John Ramey at http://johnramey.net/blog/2010/12/28/converting-a-string-to-a-variable-name-on-the-fly-and-vice-versa-in-r/
}
\examples{
x <- 42
varname2string(x)
# [1] "x"

}
