% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strings.R
\name{vector.to.single.string}
\alias{vector.to.single.string}
\title{vector.to.single.string}
\usage{
vector.to.single.string(myvector, cr = 1)
}
\arguments{
\item{myvector}{A vector of character strings}

\item{cr}{Number of line breaks after each element, defaults to 1
keyword strings
export
examples
vec <- c("Mon", "Tue", "Wed", "Thur", "Fri", "Sat", "Sun")
vector.to.single.string(vec, 2)}
}
\description{
This function takes a vector of character strings and turns it into a single string.
}

