% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strings.R
\name{vector_to_single_string_w_linebreaks}
\alias{vector_to_single_string_w_linebreaks}
\title{vector_to_single_string_w_linebreaks}
\usage{
vector_to_single_string_w_linebreaks(myvector, cr = 1)
}
\arguments{
\item{myvector}{A vector of character strings}

\item{cr}{Number of line breaks after each element, defaults to 1}
}
\description{
This function takes a vector of character strings and turns it into a single string with line breaks after each item.
}
\examples{
vec <- c("Mon", "Tue", "Wed", "Thur", "Fri", "Sat", "Sun")
vector_to_single_string(vec, 2)
}
\keyword{strings}

