% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_textfile}
\alias{write_textfile}
\title{write_textfile}
\usage{
write_textfile(myvar, myfilename)
}
\arguments{
\item{myvar}{Variable that holds character string data}

\item{myfilename}{string Name of the file}
}
\description{
Write a variable with one or more character strings to a text file taken from this stackoverflow answer http://stackoverflow.com/a/2470277/718150
}
\examples{
\dontrun{
 write_textfile(mydata, "myfile.txt")
}
}
