% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ni.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Base R has %in% to see if one object is contained within another object.
Why not %ni% to see if one object is NOT contained within another?
Idea from stackoverflow
http://stackoverflow.com/questions/11302985/match-with-negation}
\usage{
... \%ni\% NA
}
\arguments{
\item{myobject}{(a vector)}
}
\description{
Base R has %in% to see if one object is contained within another object.
Why not %ni% to see if one object is NOT contained within another?
Idea from stackoverflow
http://stackoverflow.com/questions/11302985/match-with-negation
}
\examples{
5 \%ni\% c(2,4,6)
# TRUE
}
\keyword{function}

