% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removecolsAllNA.R
\name{removecolsAllNA}
\alias{removecolsAllNA}
\title{removecolsAllNA}
\usage{
removecolsAllNA(mydf)
}
\arguments{
\item{mydf}{- data frame}
}
\value{
data frame
}
\description{
This function removes "empty" data frame columns where all the values are NA. It is a wrapper for this excellent StackOverflow answer http://stackoverflow.com/a/15968937/718150
}
\examples{
Cat <- c("CatA", "CatB", "CatC")
Val1 <- c(4, 6, 5)
Val2 <- c(NA, NA, NA)
Val3 <- c(7, NA, 2)
thedf <- data.frame(Cat, Val1, Val2, Val3)
newdf <- removecolsAllNA(thedf)

}
